initConnection at children only; renamed file
