+++ SignalRWebPack/src/js/Hubs/Base/BaseUIListenHub.ts
@@ -3,7 +3,10 @@
 export abstract class BaseUIListenHub extends BasicHub {
     constructor(signalUrl: string, baseUrl: string) {
         super(signalUrl, `${baseUrl}UI`);
-        this.initUIElements();
     }
     protected abstract initUIElements();
+    protected initConnection(): void {
+        super.initConnection();
+        this.initUIElements();
+    }
 }
